refactor: move article-add-modal out of editor
